# backend/internal/dashboard/data/openapi.json
{
  "components": {
    "schemas": {
      "AuthStatusResponse": {
        "properties": {
          "authenticated": {
            "type": "boolean"
          },
          "has_password": {
            "type": "boolean"
          },
          "is_default_admin_token": {
            "type": "boolean"
          },
          "require_login": {
            "type": "boolean"
          }
        },
        "required": [
          "authenticated",
          "has_password",
          "is_default_admin_token",
          "require_login"
        ],
        "type": "object"
      },
      "ChangePasswordRequest": {
        "properties": {
          "current_password": {
            "type": "string"
          },
          "new_password": {
            "type": "string"
          }
        },
        "required": [
          "current_password",
          "new_password"
        ],
        "type": "object"
      },
      "ChangePasswordResponse": {
        "properties": {
          "message": {
            "type": "string"
          },
          "ok": {
            "type": "boolean"
          }
        },
        "required": [
          "message",
          "ok"
        ],
        "type": "object"
      },
...
      "PageStateResponse": {
        "properties": {
          "data": {}
        },
        "required": [
          "data"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "scheme": "bearer",
        "type": "http"
      },
      "cookieAuth": {
        "in": "cookie",
        "name": "fb_admin",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Dashboard admin surface. Generated by backend/cmd/openapi-emit from admin_manifest.json + dashboard.AdminAPIPaths — do not edit by hand.",
    "title": "freebucks-proxy admin API",
    "version": "dev"
  },
  "openapi": "3.0.3",
  "paths": {
    "/admin": {
      "get": {
        "operationId": "spaRoot",
        "responses": {
          "200": {
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "SPA shell (HTML)"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "summary": "SPA shell (HTML)",
        "tags": [
          "views"
        ]
      }
    },
    "/admin/": {
      "get": {
        "operationId": "spaRootSlash",
        "responses": {
          "200": {
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "SPA shell (HTML)"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "summary": "SPA shell (HTML)",
        "tags": [
          "views"
        ]
      }
    },
    "/admin/api/auth/status": {
      "get": {
        "operationId": "getAuthStatus",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthStatusResponse"
                }
              }
            },
            "description": "Dashboard auth state (login mode, factory-default check)"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "summary": "Dashboard auth state (login mode, factory-default check)",
        "tags": [
          "auth"
        ]
      }
    },
    "/admin/api/change-password": {
      "post": {
        "operationId": "changePassword",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChangePasswordResponse"
                }
              }
            },
            "description": "Rotate the admin dashboard password"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "summary": "Rotate the admin dashboard password",
        "tags": [
          "auth"
        ]
      }
    },
    "/admin/api/config": {
      "get": {
        "operationId": "getConfig",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/configData"
                }
              }
            },
            "description": "Effective config (sensitive: raw .env read)"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "summary": "Effective config (sensitive: raw .env read)",
        "tags": [
          "settings"
        ]
      }
    },
    "/admin/api/config/meta": {
      "get": {
        "operationId": "getConfigMeta",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/getConfigMetaResponse"
                }
              }
            },
            "description": "Configuration catalog (the settings form schema)"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "summary": "Configuration catalog (the settings form schema)",
        "tags": [
          "settings"
        ]
      }
    },
    "/admin/api/events": {
      "get": {
        "operationId": "events",
        "responses": {
          "200": {
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Server-sent event stream (dashboard live updates)"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "summary": "Server-sent event stream (dashboard live updates)",
        "tags": [
          "views"
        ]
      }
    },
    "/admin/api/logs": {
      "get": {
        "operationId": "getLogs",
        "parameters": [
          {
            "description": "Log level filter",
            "in": "query",
            "name": "level",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Message substring filter",
            "in": "query",
            "name": "msg",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "View window override (Go duration, clamped to 1m..168h)",
            "in": "query",
            "name": "window",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/logsData"
                }
              }
            },
            "description": "Recent log ring view model"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "summary": "Recent log ring view model",
        "tags": [
          "views"
        ]
      }
    },
    "/admin/api/logs/export": {
      "get": {
        "operationId": "exportLogs",
        "parameters": [
          {
            "description": "Unix-millis lower bound (default 0, clamped to 168h)",
            "in": "query",
            "name": "since",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Unix-millis upper bound (default now)",
            "in": "query",
            "name": "until",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/logsExportDoc"
                }
              }
            },
            "description": "Streamed versioned export of request_records plus log_entries"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "summary": "Streamed versioned export of request_records plus log_entries",
        "tags": [
          "views"
        ]
      }
    },
    "/admin/api/logs/history": {
      "get": {
        "operationId": "getLogsHistory",
        "parameters": [
          {
            "description": "Log level filter",
            "in": "query",
            "name": "level",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Message substring filter",
            "in": "query",
            "name": "msg",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Request id filter",
            "in": "query",
            "name": "req_id",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Unix-millis lower bound",
            "in": "query",
            "name": "since",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Unix-millis upper bound",
            "in": "query",
            "name": "until",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Max records (default 500)",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/logsHistoryData"
                }
              }
            },
            "description": "Persisted log records"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "summary": "Persisted log records",
        "tags": [
          "views"
        ]
      }
    },
    "/admin/api/logs/import": {
      "post": {
        "operationId": "importLogs",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/logsImportReport"
                }
              }
            },
            "description": "Gap-fill restore of an export document (INSERT OR IGNORE, 64MB cap)"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "summary": "Gap-fill restore of an export document (INSERT OR IGNORE, 64MB cap)",
        "tags": [
          "views"
        ]
      }
    },
    "/admin/api/logs/rollup": {
      "get": {
        "operationId": "getLogsRollup",
        "parameters": [
          {
            "description": "Unix-millis lower bound (default 0, clamped to 168h)",
            "in": "query",
            "name": "since",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Unix-millis upper bound (default now)",
            "in": "query",
            "name": "until",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Time-bucket width in millis (default 300000)",
            "in": "query",
            "name": "bucket_ms",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/logsRollupData"
                }
              }
            },
            "description": "Request-outcome rollup: counts, group-bys, error rate, TTFB percentiles, time buckets"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "summary": "Request-outcome rollup: counts, group-bys, error rate, TTFB percentiles, time buckets",
        "tags": [
          "views"
        ]
      }
    },
    "/admin/api/maturity/history": {
      "get": {
        "operationId": "getMaturityHistory",
        "parameters": [
          {
            "description": "Pool token index",
            "in": "query",
            "name": "token",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Unix-millis lower bound",
            "in": "query",
            "name": "since",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Max events (default 200)",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/maturityHistoryData"
                }
              }
            },
            "description": "Streak/standing transitions"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "summary": "Streak/standing transitions",
        "tags": [
          "views"
        ]
      }
    },
...
    "/admin/api/models": {
      "get": {
        "operationId": "getModels",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/modelsData"
                }
              }
            },
            "description": "Full model catalog view model (tiers, withdrawal and live offer state)"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "summary": "Full model catalog view model (tiers, withdrawal and live offer state)",
        "tags": [
          "views"
        ]
      }
    },
    "/admin/api/notices": {
      "get": {
        "operationId": "getNotices",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NoticesResponse"
                }
              }
            },
            "description": "Upstream announcements and live broadcasts"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "summary": "Upstream announcements and live broadcasts",
        "tags": [
          "views"
        ]
      }
    },
    "/admin/api/overview": {
      "get": {
        "operationId": "getOverview",
        "parameters": [
          {
            "description": "view=live returns the hot-poll subset (live numbers only); absent returns the full shape",
            "in": "query",
            "name": "view",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/overviewData"
                }
              }
            },
            "description": "Overview view model (pool, models, quota, upstream sync)"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "summary": "Overview view model (pool, models, quota, upstream sync)",
        "tags": [
          "views"
        ]
      }
    },
    "/admin/api/pages/{id}": {
      "get": {
        "operationId": "getPageState",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageStateResponse"
                }
              }
            },
            "description": "Per-page UI snapshot (absent row reads back empty, never 404)"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "summary": "Per-page UI snapshot (absent row reads back empty, never 404)",
        "tags": [
          "settings"
        ]
      },
      "put": {
        "operationId": "putPageState",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PageStateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultEnvelope"
                }
              }
            },
            "description": "Upsert one per-page UI snapshot (data must be a JSON object)"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "summary": "Upsert one per-page UI snapshot (data must be a JSON object)",
        "tags": [
          "settings"
        ]
      }
    },
    "/admin/api/quota/history": {
      "get": {
        "operationId": "getQuotaHistory",
        "parameters": [
          {
            "description": "Pool token index (-1 disables)",
            "in": "query",
            "name": "token",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Model id",
            "in": "query",
            "name": "model",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Unix-millis lower bound",
            "in": "query",
            "name": "since",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Max samples (default 500)",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/quotaHistoryData"
                }
              }
            },
            "description": "Per-model quota samples"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "summary": "Per-model quota samples",
        "tags": [
          "views"
        ]
      }
    },
    "/admin/api/require-login": {
      "post": {
        "operationId": "requireLogin",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequireLoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequireLoginResponse"
                }
              }
            },
            "description": "Enable/disable dashboard login (open mode is loopback-only)"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "summary": "Enable/disable dashboard login (open mode is loopback-only)",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/admin/api/settings/{key}": {
      "delete": {
        "operationId": "resetSetting",
        "parameters": [
          {
            "in": "path",
            "name": "key",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultEnvelope"
                }
              }
            },
            "description": "Drop one overlay row (falls back to .env/env/default)"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "summary": "Drop one overlay row (falls back to .env/env/default)",
        "tags": [
          "settings"
        ]
      }
    },
...
  }
}